CRUD completo

+++ src/main/java/com/walmartdws/ids/DWS/service/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.walmartdws.ids.DWS.service;
 
+import com.walmartdws.ids.DWS.exception.ResourceNotFoundException;
 import com.walmartdws.ids.DWS.model.Product;
 import com.walmartdws.ids.DWS.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +17,6 @@ public class ProductServiceImpl implements ProductService{
     @Autowired
     private ProductRepository productRepository;
 
-
     @Override
     public Product createProduct(Product product) {
         return productRepository.save(product);
@@ -28,29 +28,42 @@ public class ProductServiceImpl implements ProductService{
     }
 
     @Override
-    public Product getProductByBarcode(long id) {
-        Optional<Product> productDb = this.productRepository.findById(id);
-        return productDb.get();
+    public Product getProductByBarcode(String barcode) {
+        Optional<Product> productDb = this.productRepository.findById(barcode);
+
+        if (productDb.isPresent()) {
+            return productDb.get();
+        } else {
+            throw new ResourceNotFoundException("Record not found with id : " + barcode);
+        }
     }
 
     @Override
     public Product updateProduct(Product product) {
-        Optional<Product> productDb = this.productRepository.findById(product.getId());
+        Optional<Product> productDb = this.productRepository.findById(product.getBarcode());
 
-        Product productUpdate = productDb.get();
-        productUpdate.setBarcode(product.getBarcode());
-        productUpdate.setHeight(product.getHeight());
-        productUpdate.setLenght(product.getLenght());
-        productUpdate.setWeight(product.getWeight());
-        productUpdate.setWidth(product.getWidth());
-        productRepository.save(productUpdate);
-        return productUpdate;
+        if(productDb.isPresent()) {
+            Product productUpdate = productDb.get();
+            productUpdate.setBarcode(product.getBarcode());
+            productUpdate.setHeight(product.getHeight());
+            productUpdate.setLenght(product.getLenght());
+            productUpdate.setWeight(product.getWeight());
+            productUpdate.setWidth(product.getWidth());
+            productRepository.save(productUpdate);
+            return productUpdate;
+        } else {
+            throw new ResourceNotFoundException("Record not found with id : " + product.getBarcode());
+        }
 
     }
 
     @Override
-    public void deleteProduct(long id) {
-        Optional<Product> productDb = this.productRepository.findById(id);
-        this.productRepository.delete(productDb.get());
+    public void deleteProduct(String barcode) {
+        Optional<Product> productDb = this.productRepository.findById(barcode);
+        if (productDb.isPresent()) {
+            this.productRepository.delete(productDb.get());
+        } else {
+            throw new ResourceNotFoundException("Record not found with id : " + barcode);
+        }
     }
 }
